Made some changes
segmented the html output for easier readability

printed the w's for each step, along with the sum of w's, in the html.

made the spaes between tables smaller, but made tables wider.

opened and closed the file once through all the code, instead of wrapping it in a width column each time it is needed.

replaced i = 1 with i = 0 in html printing code